Fix to #2074 - allow users to opt-out from using null comparison expansion
Added UseRelationalNullComparisonsQueryAnnotation that can be used to annotate the query with information that it should not expand null comparisons.

If relationl null comparisons are required, we annotate the predicate with NonNullableExpression, so that null semantics expansion is skipped during sql gen (since we don't have access to query annotation at that point, the information must flow in the query itself).
Because of that we can only reduce the predicate in the query generation, after the second null semantics has been (potentially) applied. Before we it was possible to do it safely in CompositeRelationalExpressionTreeVisitor, but now the reducible information is needed is sql gen.

Also fix a minor bug around Annotations - if annotations were passed into the query via AnnotateQuery (FromSql, AsNoTracking, etc), we would not be using the name of lambda parameter as alias for the table, but rather use a first letter of entity type for that table. This was happening because Relinq created a subquery from the AnnotateQuery method call, that we later on optimize, but during the process we were not preserving the ItemName of the outer query. Since the AsNoTracking didn't have a lambda, Relinq couldn't figure out the proper ItemName and was using a generated name. Fix is to preserve the ItemName of the outer query and use it, instead of the inner query name, if the latter is generated (and former is not)

CR: Andrew, Micah, Diego, Arthur.